requirements.txt
# Core Telegram Bot
python-telegram-bot==20.7

# PSD Processing
psd-tools==1.14.3

# Image Processing
Pillow==10.3.0
rembg[cpu]==2.0.57

# Environment Management
python-dotenv==1.0.0

# Additional utilities for better performance
numpy==1.26.2
requests==2.31.0
